bwa align now checks if sam alignment file already exists

+++ vcaller.py
@@ -56,13 +56,16 @@ def align_bwa(output, nthreads, reference, read1, read2):
         run(['bwa', 'index', reference])
 
     # Align input sequences to the reference genome
-    click.echo('Aligning read(s) against the reference genome...')  # find way to make message fancier with custom names
-    align_args = ['bwa', 'mem', '-M', '-t', nthreads, reference, read1]
-    if read2 is not None:
-        align_args += read2
     sam_output = ''.join(output.split('.')[:-1])+'.sam'
-    with open(sam_output, "w") as align_out:
-        run(align_args, stdout=align_out)
+    if check_existence([sam_output]):
+        click.echo('Aligned SAM read file already exists!')
+    else:
+        click.echo('Aligning read(s) against the reference genome...')  # find way to make message fancier with custom names
+        align_args = ['bwa', 'mem', '-M', '-t', nthreads, reference, read1]
+        if read2 is not None:
+            align_args += read2
+        with open(sam_output, "w") as align_out:
+            run(align_args, stdout=align_out)
 
     # Sort and convert to BAM
     click.echo('Sorting and converting to BAM...')
